Add files via upload
Zachary's Karate Club is a well-known social network that represents the relationships between members of a university karate club. A conflict arose between the administrator and the instructor of the club, leading to a split into two groups. The goal of this project is to use community detection algorithms to predict this division based on the network structure and compare the predictions with the actual division of the club members after the split.